+++ roles/brew/files/install_brew.sh
@@ -1,0 +1,5 @@
+#!/usr/bin/env bash
+
+set -e
+
+curl -fsSL https://raw.githubusercontent.com/Homebrew/install/master/install.sh | bash